Fix issue with npm auth when testing packaging (#1339)
Closes #1332 

The `test-packaging.sh` script publishes packages to a local npm repo.
This uses anonymous auth, but the npm client still wants to see an auth
token in its local configuration (it wants you to run `npm adduser
<url>`). This is annoying as the auth token isn't actually used.

The script now adds a dummy value to `~/.npmrc` in order to make npm
happy.

See
https://github.com/verdaccio/verdaccio/issues/212#issuecomment-308578500