Remove React dependencies from PCD packages, use PCD UI packages (#1353)
Closes #1324 

This PR moves all of the card rendering functionality out of the
individual PCD packages, and in to matching "PCD UI" packages.
`passport-client` now uses these for rendering instead of the PCD
packages.

This enables downstream developers who do not want
React/styled-components dependencies in their builds when only using the
non-UI functionality of the PCD packages.

The important changes are in
https://github.com/proofcarryingdata/zupass/pull/1353/files#diff-7cfc9f40a18701bfba3cdc0782f56675bba378ea59e86b323b037ea0e660811d
- removal of `renderCardBody` from `PCDPackage`, and the creation of the
`PCDUI` interface, which is where `renderCardBody` has moved to. I've
also split the header rendering feature out of `renderCardBody`, so
`PCDUI` has a `getHeader` method which does the same thing that passing
`returnHeader` into `renderCardBody` used to do.

All of the other changes to packages are just implementations of the new
interfaces, splitting the React rendering code out into new packages.

Changes to `passport-client` are minimal, only what is needed to use the
new `PCDUI` interface.

---------

Co-authored-by: Ivan Chub <ichub@users.noreply.github.com>