Implement merge for sync (#1360)
[Updated after force push] Adds merge functionality to sync, filling in
the stub added in a previous diff. Major pieces included:

- The E2EESync implementation uses the underlying merge of PCDs and
subscriptions, and adds some stats accounting I hope to be useful for
logging.
- There are TODOs for some pieces I think might want to be better,
including explaining the goals and assumptions of the merge algorithm.
- The FeedsSubscritpionManager merge is implemented and tested as per
the previous revision of this PR.
- The app-level changes around subscriptions motivated by my wanting to
make sure it's okay to replace the FeedsSubscriptionManager with a new
one at merge time. Different parts of the app have different assumptions
about this. I've done my best to make it consistent.

This deserves more testing, but I did some manual testing creating
conflicts with both PCDs and subscriptions, and I saw the expected
results in both data and stats. Those expectations are met both in the
positive (no objects are ever lost) and the negative (removed objects
come back if involved in a merge.

Fixes #641
Closes #1342