Increment sync API version (#1375)
Closes #1343

This is a quck-and-dirty way to ensure that older clients which don't
yet have our sync fixes won't be able to upload or download, and thus
won't be able to lose more PCDs in conflicts.

Someday we might want to have an explicit updating scheme which prompts
the user to reload the page, but this seems to work okay for now.