Avoid looping upload failures (#1376)
This is a minimum-viable fix for #1374. The key is ensuring that a
failed upload doesn't cause a call to update() if there are no changes
to make.

It's a bit messy, and has some races if there are multiple AppState
updates in flight at once when detecting a conflict, but it solves the
immediate problem.

An alternative I'd consider with more time would be to explicitly track
the most recent upload hash attempted, and never retry an upload until
the hash changes, or the next regular poll job occurs (using an
`extraUploadRequested` variable similar to what's used for downloads and
feeds). That would change the somewhat-desirable quick retry in the case
of transient network errors. Really what's needed is a proper retry loop
with exponential backoff, but that's complicated to implement in our
async sync framework.

Fixes #1374