Robust lookup for semaphore identity PCD (#1382)
Replace all the places where semaphore identity PCD is looked up to no
longer assume it's the first one. Instead, `self.commitment` is used to
find a matching PCD. This removes some flaky assumptions, and sets us up
for supporting multiple Semaphore identities in future. The state
validator also gets an update so it can use this lookup method normally,
but also have a fallback so it report correctly when the User or
Identity are missing.

Along the way, I ran into a case which made me finally understand Ivan's
suggestion about using Wrapper<T> in React hooks. I'm interested to know
if the pattern I captured in `usePCDCollectionWithWrapper()` seems
correct and clear. It seemed to work in my limited test case.