Turn off import/no-unresolved in VSCode
ESLint complains that it can't resolve TypeScript dependencies when
it can't find the .d.ts files, even for dependencies within the same
monorepo. This is annoying as these don't exist out-of-the-box, meaning
that ESLint fills VSCode up with red squiggly lines on a fresh checkout.

This turns the rule off in VSCode. `yarn lint` will still detect the
errors, and so will notice missing declarations when run from the command
line or in CI.