Prevent recurrence of `generic-pool` issue (#1400)
I frequently see build failures for `passport-server` and
`consumer-server` relating to the `generic-pool` types:
```
passport-server:build: error TS2688: Cannot find type definition file for 'generic-pool'.
passport-server:build:   The file is in the program because:
passport-server:build:     Entry point for implicit type library 'generic-pool'
passport-server:build: error Command failed with exit code 2.
```

The issue is with `@types/generic-pool`, which is an empty stub library.
Because it is empty, it does not contain any types, and this is why we
should not depend on it - it is also unnecessary since `generic-pool`
has its own type declarations anyway. I previously removed our direct
dependency on it, but it is still being pulled in as a transitive
dependency:
```
[1/4] 🤔  Why do we have the module "@types/generic-pool"...?
[2/4] 🚚  Initialising dependency graph...
[3/4] 🔍  Finding dependency...
[4/4] 🚡  Calculating file sizes...
=> Found "@types/generic-pool@3.8.1"
info Reasons this module exists
   - "_project_#consumer-server#@opentelemetry#auto-instrumentations-node#@opentelemetry#instrumentation-generic-pool" depends on it
   - Hoisted from "_project_#consumer-server#@opentelemetry#auto-instrumentations-node#@opentelemetry#instrumentation-generic-pool#@types#generic-pool"
```

[Because TypeScript will, by default, include all `@types` packages in
the
build](https://stackoverflow.com/questions/72734737/tsconfig-json-shows-error-entry-point-for-implicit-type-library-glob#comment136369153_75934456),
it includes this broken package by default. The intermittency in the
issue seems to be because sometimes TypeScript includes the
`generic-pool` package first, and so skips `@types/generic-pool`,
although I can't confirm this.

The solution is to tell TypeScript to only include specific `@types`
packages, by using the `types` field in `tsconfig.json`. I've modified
`passport-server` and `consumer-server`'s `tsconfig.json` to do this,
which seems to have resolved the issue.