Prevent import of internal files from other `@pcd` packages (#1404)
Closes #1401 

This prevents imports of the type `import { foo } from
"@pcd/package/sub/file/path"`. If `foo` is meant to be available then it
will be exported by the top-level `@pcd/package`, otherwise it is
intended only for internal use within that package.

This rule only affects imports from `@pcd/*` packages. Other packages
may have different structures that intentionally support importing
sub-package files. This is sometimes considered to be a good design
pattern as it can allow for more efficient elimination of un-needed code
from a package, if only a specific part of the package is required.