Fail the build if installation modifies the yarn lockfile (#1405)
Closes #1333 

Previously we used `yarn install --frozen-lockfile`, but this does not
seem to produce the desired build failure in the case where a
workspace's `package.json` is out of sync with the repo's `yarn.lock`.

Now we allow `yarn install` to modify `yarn.lock`, and fail the build if
it has been modified (specifically, if `git` detects that any files have
changed in the repo).

If you look at the commit history, you can see that the previous commit
had a deliberate failure, caused by adding something to
`apps/passport-client/package.json` and not committing the updated
`yarn.lock`. See the failed Github action here:
https://github.com/proofcarryingdata/zupass/actions/runs/7491039865