Remove dynamic require (#1406)
Closes #1383.

As per #1383, using `@pcd/util/getRandomValues()` from a NextJS project
was causing the following error to appear in the terminal:
```
[ERROR] Error: Dynamic require of "crypto" is not supported
```

This was traced back to a dynamic require in
`packages/lib/util/src/CryptoHelpers.ts`:
```typescript
      // eslint-disable-next-line @typescript-eslint/no-var-requires
      const { webcrypto } = require("node:crypto");
```

Replacing the `require` with an `import` solves the problem. This was
tested by repackaging `@pcd/util` locally and testing it with the NextJS
project in which the error was occurring. The version with `import`
works.

I also tested this in our repo by testing some code that uses this code
path - in `passport-client` the `getRandomValues()` function is called
as part of encrypting the e2ee sync blob, which continues to work as
expected.

A similar use of `require()` in `passport-crypto` was also changed to
`import`.

## Why was this happening?

It would be nice to understand the principle behind this error. Chains
of transpilation, TS -> JS and (in some cases) ESM -> CJS can be tricky
to understand. A general observation I've made is that mixing ESM-style
`import`s and CJS-style `require`s is a bad idea, and since TypeScript
uses `import` by default, it's best to stick to using `import`
exclusively. There may be some configuration of transpiler and bundler
in which it's safe to mix them, but relying on this seems unwise.

I've enabled an eslint rule to prevent future use of `require()`.