Use `RateLimitService` for all rate-limiting (#1399)
Closes #1379 

We currently have two separate rate-limiting mechanisms, one for account
resets and a generic rate limiting service which is used for email token
generation and verification. Account resets are limited to 5 per day,
whereas token generation and verification are limited to 10 per hour.
Account resets are tracked via a field in the `users` table, whereas
token generation/verification is tracked in the `rate_limit_buckets`
table.

This PR makes the `RateLimitService` more flexible, allowing the
specification of a `maxActions` and `timePeriod` option for each rate
limit type. This allows the `RateLimitService` to take over the job
performed by the custom code managing rate-limiting for account resets.
The call to the `RateLimitService` is a drop-in replacement to the check
being done during account reset, and allows for all of the
account-reset-specific rate limiting code to be removed.

**UPDATE**: the entire rate limiting logic is now implemented in
TypeScript. The DB is only used for persistence of rate limit state.

**FURTHER UPDATE**: the rate limiting logic is back in SQL again, though
this time just an `INSERT...ON CONFLICT DO UPDATE`.