Attach vscode debugger to passport-server (#1416)
Beginnings of addressing #569

This tells `passport-server` to listen for debugger requests on port
4321, and provides a VSCode `launch.json` config to attach a debugger on
that port. I've tested it and it works, allows us to set breakpoints,
step through code, etc.

I investigated attaching a debugger to tests, but this seems a bit
harder: the `ts-mocha` test process is fairly short-lived, and rather
than attaching to it, we would want to launch it from VSCode. As far as
I can tell, this might require having a configuration for each
workspace's tests, rather than a generic one for the entire test suite.