Fix build script (#1418)
The `scripts/build.sh` script had an incorrect hashbang at the
beginning. The script works in environments with reasonably modern
shells, but doesn't work in Github actions, where `sh` is the default
shell. The failure case is that build caching is disabled, which is
safe, but results in slower CI builds than necessary.

I've also tested this fix by deploying to `ivan-staging` on Render,
where it correctly uses the environment variable to disable build
caching.