Enable mandatory return types in eslint (#1417)
Closes #553 (unless we want to add more rules?)

In retrospect, it might have made sense to loosen the
`explicit-function-return-type` rule by allowing anonymous function
expressions to not specify return types. This is because front-end code
makes heavy use of small, throw-away functions whose return values
either don't matter, or are trivially inferred from their `return`
statements.

Having said that, linting now passes with the strictest version of the
rule in place, and we're free to relax it if we want to.


[https://typescript-eslint.io/linting/configs#strict](https://typescript-eslint.io/linting/configs#strict)
might be worth considering if we want to tighten rules up further.