Generic issuance checkin (#1458)
Implements check-in for both Pretix and Lemonade.

Some caveats:
- The Lemonade data model does not seem to include the fact that a
ticket has been checked in, so although it may be possible to check a
ticket in, this will not be reflected in the data we fetch from Lemonade
- Scanning of a ticket by a non-superuser will take them to a fairly
uninformative screen telling them that they do not have permission to
check the ticket in. We could implement a nicer screen for this "social
scanning" use-case, but I have not implemented the server-side functions
that would be necessary for this. It would be nice to think about the
"registry of known tickets" question before deciding what to do here.