Tracing for generic issuance (#1472)
Closes #1470

Adds Honeycomb tracing to the generic issuance service, and pipeline
implementations, covering:

- CRUD operations on pipelines
- Loading of pipeline definitions
- Loading of external data into pipelines
- Issuance of ticket PCDs
- Pre-checks for check-in
- Check-in

It's possible that we could be doing a bit more logging around errors,
but in this PR I've only added Honeycomb tracing. Aside from `setError`,
we don't seem to record a lot of data about errors in Honeycomb in the
rest of our code.