Simplify adding a subscription with a deep-link (#1476)
Closes #1474 

This makes the following changes:

If you hit `http://<zupass-client>/#/add-subscription` with a `url=?`
parameter, this will be fetched on page load. If it's a valid feed
listing URL, the available feeds will be shown with a one-click
subscribe option.

If there is an `email=?` parameter, this is taken as a "suggested
email". If this does not equal the current user's email address, they
will be warned that the feed might not be intended to be used with that
email, and they should either contact the issuer or sign up for another
Zupass account with the email the feed issuer expects:

<img width="428" alt="image"
src="https://github.com/proofcarryingdata/zupass/assets/20022/bed9838e-f3fb-43d5-8bad-80058d5b1c0b">


If the user is not logged in, then they will be taken to a login form
with the suggested email pre-filled. On completing login/registration,
they will be taken to the "add subscription" flow as described above.