Initialize frogcrypto test feeds with the exact same time value (#1485)
Fixes #1444 

The issue here is that the tests assume that all of the feeds were
initialized at the same, and thus have the same `activeUntil` value.
However, this was done with a series of sequential calls to
`Date.now()`, and sometimes a millisecond elapses between these calls
and they give different values. This is fixed by calling `Date.now()`
once and using that value to initialize all of the feeds.