Upgrade Semaphore libraries (#1515)
Upgrading our Semaphore dependencies to the latest releases of the v3
series. (v4 is a major incompatible upgrade which will have to come
separately later.)

Backward-compatibiliy tests introduced recently should ensure this isn't
breaking the ability to parse any of the PCDs already in the wild. This
has a negligible impact on bundle size (it does pull in 28K for
zk-kit/groth16 which we aren't using ourselves anymore). I didn't test
performance impact.