Target ES2020 (#1533)
Upgrading language version after discussion on Discord. The primary
motivating feature at the moment is bigint literals, so I found a few
places in code where they could be used to replace BigInt(N), and
applied those changes to ensure it works.
The Buffer import change is unrelated to ES2020, but is a followup to
the same discussion on Discord.

This seems to work for me locally. I tested with Consumer Client
requesting a Semaphore Signature, which involves a hash calculation
which now uses a bigint literal.

I'm not sure how to fully validate the browser compatibility impact of
this. Once this is on staging, an important test will be to try it out
in Telegram, which is probably the most limited browser we need to
support.