Cache issuance in Lemonade and Pretix pipelines (#1530)
Direct copy of the issuance caching mechanism from the original issuance
service, with the exception that the signing key is also included in the
cache key. This is to cover the (unlikely but possible in future)
scenario in which two pipelines issue identical tickets but with
different signing keys.

There is substantial code duplication between the two pipelines, and if
we create additional pipelines which also issue tickets then we may want
to refactor this into a `TicketIssuer` class which takes in
`ITicketData` from the pipeline and handles all of the `EdDSATicketPCD`
proving, and caching of the result.