Longer feed credential expiry (#1540)
Feed credentials are cached for an hour on the client side, to prevent
regenerating them too often.

However, feed credentials also expire after an hour. This means that,
close to the end of the hour, the cache is serving credentials that are
close to expiry. With some clock skew and/or latency, we see expired
credentials occasionally being used.

This PR makes the expiry time 80 minutes instead of 60, meaning that
credentials will expire from the client-side cache 20 minutes before
they become useless.