Manually-configured tickets (#1543)
Closes [OXP-28](https://linear.app/0xparc-pcd/issue/0XP-28)

Manually-configured tickets can be added to both Lemonade and Pretix
pipelines, by adding the following section to the Pipeline definition
options:
```json
{
  "manualTickets": [
    {
      "id": "b3bb20f5-7aed-4963-baaf-a8c6010739b0",
      "eventId": "4a2e7e30-8f4e-462a-a13b-6c69c14e9ea9",
      "productId": "977e5b25-be0a-4058-8eea-9e5922b9eb3d",
      "attendeeEmail": "manual@example.com",
      "attendeeName": "Manual Test"
    }
  ]
}
```

The ID is a UUID which the admin user will need to generate. The event
ID and product ID *must* match an event ID/product ID pair in the
pipeline definition, and will fail to validate if they do not.

If the product ID is a superuser product, then holders of these tickets
can perform check-ins. However, manually-configured tickets can not
themselves be checked in (a later PR will add support for this).