Check-in for manual tickets (#1544)
Closes https://linear.app/0xparc-pcd/issue/0XP-39

Allows manually-specified tickets to be checked in.

This is achieved by adding a database table in which check-ins are
recorded. On check-in, the pipeline determines if the ticket to be
checked in is a remote ticket (from Pretix or Lemonade) or a manual
ticket. If the former, it calls the remote API to check the ticket in,
and if the latter it persists the check-in to the DB.