Generic issuance semaphore groups (#1549)
closes #1475

Introduces a few new things:

- A table of "consumers" of generic issuance feeds, storing their emails
and Semaphore IDs
- A table of Semaphore group history, identical to the old Semaphore
group history but with added pipeline ID
- PipelineConsumerDB for managing the pipeline consumer data
- PipelineSemaphoreHistoryDB for managing Semaphore group history
- SemaphoreGroupProvider, which provides generic features that can be
used by pipelines to create Semaphore group histories. Pipelines are
responsible for telling their SemaphoreGroupProvider about the set of
tickets they have whenever it changes, but otherwise the Pipeline
classes do not need to know anything specific about Semaphore groups.
This should avoid code duplication when we need to add this
functionality to other pipelines.
- SemaphoreGroupCapability, so the Generic Issuance service knows which
Pipelines offer Semaphore groups and can invoke the required
functionality
- Various API classes, routes, handlers, and so on
- New pipeline configuration options for semaphore groups, see example
in the generic issuance tests

So far I've only added this to the Lemonade pipeline, but extending it
to others should be easy.