Fetch email PCD as part of account setup (#1568)
Closes https://linear.app/0xparc-pcd/issue/0XP-168

Because the email PCD is a requirement for some default subscriptions to
work, we need to ensure that the user has an email PCD at an earlier
stage of login for first use. This PR causes the email PCD to be fetched
as part of account setup.

I have been testing this locally by adding a manual ticket for a new
email in Podbox, then signing into Zupass locally using that email, and
observing that the tickets for the Podbox feed appear after first
login/registration without requiring a refresh or any other interaction.