Update Semaphore groups whenever consumers change (#1566)
Fixes an issue we encountered when turning on Semaphore groups for Edge
City.

Specifically:

Consumer records are only kept if Semaphore groups are enabled, so when
Semaphore groups are first enabled, the consumer table is empty. It is
gradually populated as users request tickets. Since hitting a prove
screen causes a full sync, we can rely on the table being populated even
for prove screen interactions.

However, Semaphore groups which are based on the consumer table were
only updated asynchronously, on a timer. This meant that users who were
coming via Zupoll and hitting Zupass for the first time since Semaphore
groups were enabled would not be present in the group immediately.

This PR implements a synchronous update to the Semaphore groups every
time the consumer table changes, which would eliminate that problem.