License tweaks based on dependencies (#1572)
I'm trying to be careful and intentional about where GPL licensing does
and doesn't have to apply to POD code based on dependencies. I've
tweaked my own best-practice here after a talk with @gubsheep. Changes
in this PR:

@pcd/util has only MIT-licensed dependencies, so doesn't have to be GPL.
Changing it to MIT will free @pcd/pod to do the same.

Meanwhile, the perftest tool depends on lots of PCD types which need GPL
dependencies (notably circomlib, circomlibjs), so it needs to be marked
GPL.