Remove duplicate call to rate-limited function in account reset flow (#1575)
Closes https://linear.app/0xparc-pcd/issue/0XP-156

The account reset flow is rate-limited to 5 account resets per day. The
"verify token" API endpoint is limited to 10 requests per hour.

If the user performs 5 sequential account resets, we would expect them
to be rate-limited by the account reset limit, and to receive the
appropriate error message.

However, the account reset flow was involves calling the
`checkTokenCorrect` method on the `EmailTokenService` twice - once in
the "verify token" API endpoint, and again from within `handleNewUser`.
Because of this, the user will exhaust the bucket for token verification
before they exhaust the account reset bucket, and will not see the
expected error message.

The intent behind the limit on `checkTokenCorrect` was intended to limit
access to the verify token API, and not to limit new user creation. It
is probably a mistake that the rate limiting is applied to
`checkTokenCorrect`, rather than being applied to the endpoint itself.

This PR removes the rate limit on `EmailTokenService.checkTokenCorrect`,
and instead applies one to `handleVerifyToken`. `handleNewUser` now no
longer triggers the rate limiter when it calls `checkTokenCorrect`.

This means that each account reset only increments the rate limit for
token verification once, and therefore does not exhaust the token
verification rate limit before the account reset limit, meaning that the
user sees the correct error message when they hit the account reset
limit:


![image](https://github.com/proofcarryingdata/zupass/assets/20022/75a5c3c1-da60-493c-9e68-b57745844eaf)