Fix flaky Semaphore group test (#1604)
Closes
https://linear.app/0xparc-pcd/issue/0XP-252/fix-flaky-semaphore-service-testimplementation#comment-2322d5eb

In the Devconnect tests, we log some users in and then test to see if
those user's commitments show up in the expected Semaphore groups.
However, the `UserService` only *schedules* an asynchronous reload of
the Semaphore groups, and sometimes this does not occur before the
subsequent test runs. This means that sometimes the commitments are not
in the expected group.

In this PR I've added an explicit call to the Semaphore service to
reload groups before testing the group membership.