Fetch Email PCD subscription first if the user has no Email PCD (#1608)
Closes
https://linear.app/0xparc-pcd/issue/0XP-205/ensure-that-zupass-users-who-dont-get-their-email-pcd-initially-have

Since we now have default subscriptions that use Email PCDs as
credentials, it is important that the Email PCD is present before
polling these subscriptions. Previously we implemented a feature to poll
the Email PCD subscription during account creation, which should ensure
that the Email PCD is present before the first attempt to poll other
subscriptions.

However, this could still fail due to connectivity issues or downtime.
Or, the user could delete their Email PCD.

This means that the next attempt to fetch subscriptions using the Email
PCD as a credential will fail. Since each periodic polling of
subscriptions will also poll the Email PCD subscription, the Email PCD
ought to be restored; however, since this happens in parallel with the
polling of other subscriptions, the user will have to wait one minute
for the *next* poll, or refresh Zupass.

In this PR, the batch polling function now checks to see if the Email
PCD is present before polling and, if not, separately polls the Email
PCD subscription first, before proceeding to poll all other
subscriptions.

This cannot fully protect us against connectivity or downtime issues,
but it does ensure that if the feed is available then only a single
attempt to poll subscriptions is necessary to get to the best possible
state, in which the user has an Email PCD and the subscriptions which
depend on the Email PCD can be polled.

I have tested this by:
- Disabling the fetch of the Email PCD during account creation
- Deleting the Email PCD and reloading the app

In both cases, the Email PCD is restored on the next subscription poll,
and the other feeds are successfully polled.